move linalg funcs to tools.linalg
