+++ sm2/tools/linalg.py
@@ -5,7 +5,7 @@ changes:
 The only changes are that additional results are returned
 
 """
-# TODO: This is only needed (so far) by var_model.  Maybe avoid porting?
+from six.moves import reduce
 
 import numpy as np
 import scipy.linalg
@@ -17,7 +17,6 @@ _array_precision = {'f': 0, 'd': 1, 'F': 0, 'D': 1}
 
 
 # Linear Least Squares
-
 def lstsq(a, b, cond=None, overwrite_a=0, overwrite_b=0):  # pragma: no cover
     raise NotImplementedError("lstsq not ported from upstream")
 
@@ -53,55 +52,6 @@ def logdet_symm(m, check_symm=False):
 
 def stationary_solve(r, b):  # pragma: no cover
     raise NotImplementedError("stationary_solve not ported from upstream")
-
-
-# upstream this is implemented in regression.mixed_linear_model
-def _dot(x, y):
-    """
-    Returns the dot product of the arrays, works for sparse and dense.
-    """
-    if isinstance(x, np.ndarray) and isinstance(y, np.ndarray):
-        return np.dot(x, y)
-    elif sparse.issparse(x):
-        return x.dot(y)
-    elif sparse.issparse(y):
-        return y.T.dot(x.T).T
-
-
-# upstream this is implemented in regression.mixed_linear_model
-def _multi_dot_three(A, B, C):
-    """
-    Find best ordering for three arrays and do the multiplication.
-
-    Doing in manually instead of using dynamic programing is
-    approximately 15 times faster.
-
-    From numpy, adapted to work with sparse and dense arrays.
-    """
-    # cost1 = cost((AB)C)
-    cost1 = (A.shape[0] * A.shape[1] * B.shape[1] +  # (AB)
-             A.shape[0] * B.shape[1] * C.shape[1])   # (--)C
-    # cost2 = cost((AB)C)
-    cost2 = (B.shape[0] * B.shape[1] * C.shape[1] +  # (BC)
-             A.shape[0] * A.shape[1] * C.shape[1])   # A(--)
-
-    if cost1 < cost2:
-        return _dot(_dot(A, B), C)
-    else:
-        return _dot(A, _dot(B, C))
-
-
-# upstream this is implemented in regression.mixed_linear_model
-def _dotsum(x, y):
-    """
-    Returns sum(x * y), where '*' is the pointwise product, computed
-    efficiently for dense and sparse matrices.
-    """
-    if sparse.issparse(x):
-        return x.multiply(y).sum()
-    else:
-        # This way usually avoids allocating a temporary.
-        return np.dot(x.ravel(), y.ravel())
 
 
 # upstream this is implemented in regression.mixed_linear_model._smw_solver
@@ -198,3 +148,130 @@ def smw_logdet(s, A, AtA, BI, di, B_logdet):
         qmat = qmat.todense()
     _, ld1 = np.linalg.slogdet(qmat)
     return B_logdet + ld + ld1
+
+
+# upstream this is in tools.tools
+def pinv_extended(X, rcond=1e-15):
+    """
+    Return the pinv of an array X as well as the singular values
+    used in computation.
+
+    Code adapted from numpy.
+    """
+    X = np.asarray(X)
+    X = X.conjugate()
+    u, s, vt = np.linalg.svd(X, 0)
+    s_orig = np.copy(s)
+    m = u.shape[0]
+    n = vt.shape[1]
+    cutoff = rcond * np.maximum.reduce(s)
+    for i in range(min(n, m)):
+        if s[i] > cutoff:
+            s[i] = 1. / s[i]
+        else:
+            s[i] = 0.
+    res = np.dot(np.transpose(vt), np.multiply(s[:, np.core.newaxis],
+                                               np.transpose(u)))
+    return res, s_orig
+
+
+# -------------------------------------------------------------------
+# Multiplication-like Functions
+
+def chain_dot(*arrs):
+    """
+    Returns the dot product of the given matrices.
+
+    Parameters
+    ----------
+    arrs: argument list of ndarray
+
+    Returns
+    -------
+    Dot product of all arguments.
+
+    Examples
+    --------
+    >>> import numpy as np
+    >>> from sm2.tools import chain_dot
+    >>> A = np.arange(1,13).reshape(3,4)
+    >>> B = np.arange(3,15).reshape(4,3)
+    >>> C = np.arange(5,8).reshape(3,1)
+    >>> chain_dot(A,B,C)
+    array([[1820],
+       [4300],
+       [6780]])
+    """
+    return reduce(lambda x, y: np.dot(y, x), arrs[::-1])
+
+
+def nan_dot(A, B):
+    """
+    Returns np.dot(left_matrix, right_matrix) with the convention that
+    nan * 0 = 0 and nan * x = nan if x != 0.
+
+    Parameters
+    ----------
+    A, B : np.ndarrays
+    """
+    # Find out who should be nan due to nan * nonzero
+    should_be_nan_1 = np.dot(np.isnan(A), (B != 0))
+    should_be_nan_2 = np.dot((A != 0), np.isnan(B))
+    should_be_nan = should_be_nan_1 + should_be_nan_2
+
+    # Multiply after setting all nan to 0
+    # This is what happens if there were no nan * nonzero conflicts
+    C = np.dot(np.nan_to_num(A), np.nan_to_num(B))
+
+    C[should_be_nan] = np.nan
+
+    return C
+
+
+# upstream this is implemented in regression.mixed_linear_model
+def _dotsum(x, y):
+    """
+    Returns sum(x * y), where '*' is the pointwise product, computed
+    efficiently for dense and sparse matrices.
+    """
+    if sparse.issparse(x):
+        return x.multiply(y).sum()
+    else:
+        # This way usually avoids allocating a temporary.
+        return np.dot(x.ravel(), y.ravel())
+
+
+# upstream this is implemented in regression.mixed_linear_model
+def _dot(x, y):
+    """
+    Returns the dot product of the arrays, works for sparse and dense.
+    """
+    if isinstance(x, np.ndarray) and isinstance(y, np.ndarray):
+        return np.dot(x, y)
+    elif sparse.issparse(x):
+        return x.dot(y)
+    elif sparse.issparse(y):
+        return y.T.dot(x.T).T
+
+
+# upstream this is implemented in regression.mixed_linear_model
+def _multi_dot_three(A, B, C):
+    """
+    Find best ordering for three arrays and do the multiplication.
+
+    Doing in manually instead of using dynamic programing is
+    approximately 15 times faster.
+
+    From numpy, adapted to work with sparse and dense arrays.
+    """
+    # cost1 = cost((AB)C)
+    cost1 = (A.shape[0] * A.shape[1] * B.shape[1] +  # (AB)
+             A.shape[0] * B.shape[1] * C.shape[1])   # (--)C
+    # cost2 = cost((AB)C)
+    cost2 = (B.shape[0] * B.shape[1] * C.shape[1] +  # (BC)
+             A.shape[0] * A.shape[1] * C.shape[1])   # A(--)
+
+    if cost1 < cost2:
+        return _dot(_dot(A, B), C)
+    else:
+        return _dot(A, _dot(B, C))
